+++ crates/ibc/src/context/transfer_mod.rs
@@ -5,9 +5,8 @@ use std::collections::BTreeSet;
 use std::fmt::Debug;
 use std::rc::Rc;
 
-use namada_core::address::Address;
-use namada_core::ibc::apps::transfer::context::TokenTransferValidationContext;
-use namada_core::ibc::apps::transfer::module::{
+use ibc::apps::transfer::context::TokenTransferValidationContext;
+use ibc::apps::transfer::module::{
     on_acknowledgement_packet_execute, on_acknowledgement_packet_validate,
     on_chan_close_confirm_execute, on_chan_close_confirm_validate,
     on_chan_close_init_execute, on_chan_close_init_validate,
@@ -18,21 +17,18 @@ use namada_core::ibc::apps::transfer::module::{
     on_recv_packet_execute, on_timeout_packet_execute,
     on_timeout_packet_validate,
 };
-use namada_core::ibc::apps::transfer::types::error::TokenTransferError;
-use namada_core::ibc::apps::transfer::types::MODULE_ID_STR;
-use namada_core::ibc::core::channel::types::acknowledgement::Acknowledgement;
-use namada_core::ibc::core::channel::types::channel::{Counterparty, Order};
-use namada_core::ibc::core::channel::types::error::{
-    ChannelError, PacketError,
-};
-use namada_core::ibc::core::channel::types::packet::Packet;
-use namada_core::ibc::core::channel::types::Version;
-use namada_core::ibc::core::host::types::identifiers::{
-    ChannelId, ConnectionId, PortId,
-};
-use namada_core::ibc::core::router::module::Module;
-use namada_core::ibc::core::router::types::module::{ModuleExtras, ModuleId};
-use namada_core::ibc::primitives::Signer;
+use ibc::apps::transfer::types::error::TokenTransferError;
+use ibc::apps::transfer::types::MODULE_ID_STR;
+use ibc::core::channel::types::acknowledgement::Acknowledgement;
+use ibc::core::channel::types::channel::{Counterparty, Order};
+use ibc::core::channel::types::error::{ChannelError, PacketError};
+use ibc::core::channel::types::packet::Packet;
+use ibc::core::channel::types::Version;
+use ibc::core::host::types::identifiers::{ChannelId, ConnectionId, PortId};
+use ibc::core::router::module::Module;
+use ibc::core::router::types::module::{ModuleExtras, ModuleId};
+use ibc::primitives::Signer;
+use namada_core::address::Address;
 
 use super::common::IbcCommonContext;
 use super::token_transfer::TokenTransferContext;
@@ -351,10 +347,8 @@ fn into_packet_error(error: TokenTransferError) -> PacketError {
 pub mod testing {
     use std::str::FromStr;
 
-    use namada_core::ibc::apps::transfer::types::{
-        ack_success_b64, PORT_ID_STR,
-    };
-    use namada_core::ibc::core::channel::types::acknowledgement::AcknowledgementStatus;
+    use ibc::apps::transfer::types::{ack_success_b64, PORT_ID_STR};
+    use ibc::core::channel::types::acknowledgement::AcknowledgementStatus;
 
     use super::*;
